Standup_agent/requirements.txt
google-adk
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
python-dateutil
